requirements.txt
gevent
gunicorn
markdown2
pymysql
django==1.11.5
djangorestframework
markdown 
django-filter
monthdelta
Pillow
xlrd
xlsxwriter
djangorestframework
markdown
